+++ src/Protocols/Query36.php
@@ -20,10 +20,11 @@ class Query36
         curl_setopt(
             $lgsl_fp,
             CURLOPT_URL,
-            "https://discord.com/api/v9/invites/{$server[SParams::BASIC]['ip']}?with_counts=true"
+            "https://discord.com/api/v10/invites/{$server[SParams::BASIC]['ip']}?with_counts=true"
         );
 
         $buffer = curl_exec($lgsl_fp);
+
         $buffer = json_decode($buffer, true);
 
         if (isset($buffer['message'])) {
@@ -32,21 +33,19 @@ class Query36
         }
 
         $server[SParams::SERVER]['map'] = 'discord';
-        $server[SParams::SERVER]['name'] = $buffer['guild']['name'];
-        $server[SParams::SERVER]['players'] = $buffer['approximate_presence_count'];
-        $server[SParams::SERVER]['playersmax'] = $buffer['approximate_member_count'];
-        $server[SParams::CONVARS]['id'] = $buffer['guild']['id'];
+        $server[SParams::SERVER]['name'] = self::checkValue($buffer['guild']['name']);
+        $server[SParams::SERVER]['players'] = self::checkValue($buffer['approximate_presence_count'], 2);
+        $server[SParams::SERVER]['playersmax'] = self::checkValue($buffer['approximate_member_count'], 2);
+        $server[SParams::CONVARS]['id'] = self::checkValue($buffer['guild']['id'], 2);
 
-        if($buffer['guild']['description']) {
-            $server[SParams::CONVARS]['description'] = $buffer['guild']['description'];
-        }
+        $server[SParams::CONVARS]['description'] = self::checkValue($buffer['guild']['description']);
 
-        if($buffer['guild']['welcome_screen'] && $buffer['guild']['welcome_screen']['description']) {
+        if(isset($buffer['guild']['welcome_screen']) && $buffer['guild']['welcome_screen']['description']) {
             $server[SParams::CONVARS]['description'] = $buffer['guild']['welcome_screen']['description'];
         }
 
-        $server[SParams::CONVARS]['features'] = implode(', ', $buffer['guild']['features']);
-        $server[SParams::CONVARS]['nsfw'] = (int) $buffer['guild']['nsfw'];
+        $server[SParams::CONVARS]['features'] = implode(', ', self::checkValue($buffer['guild']['features']));
+        $server[SParams::CONVARS]['nsfw'] = self::checkValue($buffer['guild']['nsfw']);
 
         if (isset($buffer['inviter'])) {
             $server[SParams::CONVARS]['inviter'] = $buffer['inviter']['username'] . "#" . $buffer['inviter']['discriminator'];
@@ -83,5 +82,18 @@ class Query36
         }
 
         return true;
+    }
+
+    public static function checkValue($value, $type = 1)
+    {
+        if(isset($value)){
+            return $value;
+        }
+
+        switch($type)
+        {
+            case 1: { return "Value is undefined"; }
+            case 2: { return 0; }
+        }
     }
 }